// src/shared/models/open-graph/link-meta.ts
export interface LinkMeta {
  author?: string | undefined;
  date?: string | undefined;
  logo?: string | undefined;
  publisher?: string | undefined;
  description?: string | undefined;
  image?: string | undefined;
  title?: string | undefined;
  url?: string | undefined;
  video?: string | undefined;
  lang?: string | undefined;
  audio?: string | undefined;
  type?: string | undefined;
}